refactor(#1682): rewrite refs reyn.dispatch → reyn.core.dispatch (broad-#5 Phase A2)
19 files (imports + self-imports across src+tests+scripts+dogfood), zero
bare reyn.dispatch remain. class-3 clean (no source-path-reads); code-only
(no runtime dir). ruff check src tests --fix clean.
