+++ tests/test_dispatcher_wal_step_events.py
@@ -16,8 +16,8 @@ from __future__ import annotations
 import asyncio
 from typing import Any
 
+from reyn.core.dispatch import DispatchContext, dispatch_tool
 from reyn.core.events.state_log import StateLog
-from reyn.dispatch import DispatchContext, dispatch_tool
 
 # ---------------------------------------------------------------------------
 # Helpers
