+++ tests/test_memo_purity_e2e.py
@@ -22,8 +22,8 @@ import asyncio
 from pathlib import Path
 from typing import Any
 
+from reyn.core.dispatch import DispatchContext, dispatch_tool
 from reyn.core.events.state_log import StateLog
-from reyn.dispatch import DispatchContext, dispatch_tool
 from reyn.skill.skill_resume_analyzer import SkillResumeAnalyzer
 from reyn.skill.skill_snapshot import SkillSnapshot
