+++ tests/test_runtime_llm_memoization.py
@@ -19,9 +19,9 @@ from __future__ import annotations
 import asyncio
 
 import reyn.core.kernel.llm_call_recorder as runtime_mod
+from reyn.core.dispatch.dispatcher import _compute_llm_args_hash
 from reyn.core.events.state_log import StateLog
 from reyn.core.kernel.runtime import OSRuntime
-from reyn.dispatch.dispatcher import _compute_llm_args_hash
 from reyn.llm.llm import LLMCallResult
 from reyn.schemas.models import (
     Phase,
